Interaction Design Final: Website Project
This was the culminating assignment from my Interaction Design course where they taught responsive UI/UX design, human-centered design, and developed our skills in CSS & HTML. We were to create a website as a designer/coder using a creative brief from a partner in the class who served as the Creative Director (who decided the topic and style). This is a multi-page website about The Color Pink.